AthleteManagement Assignment
Allows a user to create an list of athletes and save their info. The
user can sort the list by name or by medal count. The user can delete an
athlete by specifying the athlete's name, and can clear the list. The
user can save the list, and load it later. The user can save a text
document and read it.